+++ cmd/run/run.go
@@ -422,7 +422,11 @@ func convertStringArrayToUintArray(stringArray []string) []uint {
 // The context provided to this function should be error-free, or shut down will be incomplete.
 func (s *ServerContext) telemetryConfig(config *serverconfig.Config) func() error {
 	if config.Trace.Enabled {
-		s.Logger.Info(fmt.Sprintf("🕵 tracing enabled: sampling ratio is %v and sending traces to '%s', tls: %t", config.Trace.SampleRatio, config.Trace.OTLP.Endpoint, config.Trace.OTLP.TLS.Enabled))
+		endpoint, schemeInsecure := telemetry.ParseOTLPEndpoint(config.Trace.OTLP.Endpoint)
+		configInsecure := !config.Trace.OTLP.TLS.Enabled
+		effectiveTLS := !telemetry.ResolveOTLPInsecure(configInsecure, schemeInsecure)
+
+		s.Logger.Info(fmt.Sprintf("🕵 tracing enabled: sampling ratio is %v and sending traces to '%s', tls: %t", config.Trace.SampleRatio, endpoint, effectiveTLS))
 
 		options := []telemetry.TracerOption{
 			telemetry.WithOTLPEndpoint(